+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 ### Added
 * New fixtures and integration tests
+* Support for paging layers from server version 10.0
 
 ### Changed
 * Refactored paging strategy