+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [1.1.0] - 2015-08-10
 ### Added
 * New fixtures and integration tests for paging
 * Support for paging layers from server version 10.0
@@ -66,6 +66,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Code for requesting data from FeatureServices
 * Tests on most methods
 
+[1.1.0]: https://github.com/chelm/featureservice/ompare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/chelm/featureservice/ompare/v0.2.0...v1.0.0
 [0.2.0]: https://github.com/chelm/featureservice/ompare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/chelm/featureservice/ompare/v0.0.4...v0.1.0